- Geometry::WireCoordinate() returns double now: rounding errors were relevant - added a test for the new implementation of WireIDsIntersect()

+++ Geometry/Geometry.h
@@ -181,10 +181,10 @@ namespace geo {
 				      unsigned int const TPCNo = 0,
 				      unsigned int const cstat = 0)  const; // nearest wire to input
                                                                             // world coordinates
-    float WireCoordinate(float YPos, float ZPos,
-                         unsigned int PlaneNo,
-                         unsigned int TPCNo,
-                         unsigned int cstat) const;
+    double WireCoordinate(double YPos, double ZPos,
+                          unsigned int PlaneNo,
+                          unsigned int TPCNo,
+                          unsigned int cstat) const;
 
     unsigned int       NearestWire(const double worldLoc[3],
 				   unsigned int const PlaneNo,
